update stories to arrange by components
