+++ package.json
@@ -33,7 +33,8 @@
     "ti": "yarn test:integration",
     "tr": "yarn coverage && yarn report && opn .coverage/index.html",
     "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "deploy-storybook": "storybook-to-ghpages"
   },
   "pre-push": [
     "lint",
@@ -114,6 +115,7 @@
   },
   "dependencies": {
     "@storybook/addon-backgrounds": "^4.0.0-alpha.10",
+    "@storybook/storybook-deployer": "^2.3.0",
     "lodash.get": "^4.4.2",
     "lodash.set": "^4.3.2",
     "makeup-expander": "~0.4.0",
