--- stories/components/click-count/index.marko
@@ -1,18 +1,0 @@
-class {
-  onCreate() {
-    this.state = { count:0 };
-  }
-  increment() {
-    this.state.count++;
-  }
-  onDestroy() {
-    console.log('click-count onDestroy');
-  }
-}
-
-<div.count>
-  ${state.count}
-</div>
-<button.example-button on-click('increment')>
-  Click me!
-</button>
