--- stories/components/click-count/style.css
@@ -1,9 +1,0 @@
-.count {
-  color:#70b;
-  font-size:3em;
-}
-
-.example-button {
-  font-size:1em;
-  padding:0.5em;
-}
